checkers/pokupaika/requirements.txt
pycryptodome
PyJWT
faker